Vault should have k8s-signed cert.

+++ ubuntu-utm-vms/vm-prep/install-calico-cni.sh
@@ -9,12 +9,18 @@ fi
 cp ${KUBECONFIG} /etc/cni/net.d/calico-kubeconfig
 chmod 600 /etc/cni/net.d/calico-kubeconfig
 
-cp /usr/share/host/guest/all-nodes/downloads/calico /opt/cni/bin/
-cp /usr/share/host/guest/all-nodes/downloads/calico-ipam /opt/cni/bin/
+if ! [ -f /opt/cni/bin/calico ]; then
+    cp /usr/share/host/guest/all-nodes/downloads/calico /opt/cni/bin/
+fi
+if ! [ -f /opt/cni/bin/calico-ipam ]; then
+    cp /usr/share/host/guest/all-nodes/downloads/calico-ipam /opt/cni/bin/
+fi
 
 chmod 755 /opt/cni/bin/calico*
 
-if test -f /etc/cni/net.d/*.conflist; then
-    for file in /etc/cni/net.d/*.conflist; do mv "$file" "${file/conflist/conflist.disabled}"; done
+confFiles=(/etc/cni/net.d/*.conflist)
+if [ ${#confFiles[@]} ]; then
+    # for file in /etc/cni/net.d/*.conflist; do mv "$file" "${file/conflist/conflist.disabled}"; done
+    for file in "${confFiles[@]}"; do mv "$file" "${file/conflist/conflist.disabled}"; done
 fi
 cp /usr/share/host/guest/cni/10-calico.conflist /etc/cni/net.d/
